third_party/WebKit/LayoutTests/webaudio/osc-custom-sweep-snr-expected.txt

Test Custom Oscillator with Exponential Sweep

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Exceeded SNR threshold of 83.89 dB
PASS Maximum difference below threshold of 2.22 ulp (16-bits)
PASS Number of differences between actual and expected result is less than 4450
PASS successfullyParsed is true

TEST COMPLETE
